requirements.txt
pytest
playwright
pyyaml
pytest-playwright
